Make the file tree search for the SSH executable opt-in (#1)
On non-linux build agents it is likely that the file tree search is done.
This means a serious performance hit for build agent startup
and a big security issue if you run untrusted builds.

Due to that now you have to explicitly configure the SSH executable
to 'auto' to enable the file tree scan if you really want it.